refactor(core): fold worker/index duplication into Linter module
Candidate 1 of the architecture review. The lint-scheduling chain
(initializeESLint → filter → shouldIgnoreModule → lintFiles) was mirrored
in index.ts and worker.ts. Collapse it into a deep Linter module; both
callers now act as adapters.

- linter.ts: createLinter factory, Linter interface, shouldIgnoreModule
  (exported for testing), and all private collaborators
- index.ts: main-thread adapter, holds worker/linter/currentContext
- worker.ts: worker adapter, 25 lines (was 73), message-forwarding only
- utils.ts: trimmed to getOptions
- linter.test.ts: shouldIgnoreModule branch coverage (4 cases)

Verified: type-check, vitest (4/4), biome check, tsdown build all clean.
